INIT and refractoring
