+++ .gitignore
@@ -1,0 +1,4 @@
+.idea
+.DS_Store
+/tmp/*
+result.json
