+++ Gemfile
@@ -1,0 +1,9 @@
+source 'https://rubygems.org'
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
+
+ruby '2.7.1'
+
+gem 'minitest'
+gem 'rspec-benchmark'
+gem 'ruby-prof'
+gem 'stringio'
