+++ case-study.md
@@ -12,18 +12,32 @@
 Я решил исправить эту проблему, оптимизировав эту программу.
 
 ## Формирование метрики
-Для того, чтобы понимать, дают ли мои изменения положительный эффект на быстродействие программы я придумал использовать такую метрику: *тут ваша метрика*
+Для того, чтобы понимать, дают ли мои изменения положительный эффект на быстродействие программы я сформировал 3 файла(на 1536, 3072 и 6144 строк) на которых мы будем замерять производительность. 
+Бенчмарк перед оптимизацией показал - что при увеличение данных в 2 раза - скорость обработки увеличивается чуть более чем в 3 раза:
+
+    1536-rows:  0.058431
+    3072-rows:  0.192021
+    6144-rows:  0.665419
+
 
 ## Гарантия корректности работы оптимизированной программы
 Программа поставлялась с тестом. Выполнение этого теста в фидбек-лупе позволяет не допустить изменения логики программы при оптимизации.
 
 ## Feedback-Loop
-Для того, чтобы иметь возможность быстро проверять гипотезы я выстроил эффективный `feedback-loop`, который позволил мне получать обратную связь по эффективности сделанных изменений за *время, которое у вас получилось*
+Для того, чтобы иметь возможность быстро проверять гипотезы я выстроил эффективный `feedback-loop`, 
+который позволил мне получать обратную связь по эффективности сделанных изменений за 25секунд
 
-Вот как я построил `feedback_loop`: *как вы построили feedback_loop*
+Вот как я построил `feedback_loop`: 
+- Тестирование работоспосбности программы: самое важное - чтоб все работало правильно
+- Тестирование метрик - Для защиты от потери достигнутого прогресса
+- Бенчмарк
+- Профилирование ruby prof
+
 
 ## Вникаем в детали системы, чтобы найти главные точки роста
-Для того, чтобы найти "точки роста" для оптимизации я воспользовался *инструментами, которыми вы воспользовались*
+Перед началом оптимизации - я решил отрефракторить код, разбив основные этапы программы на именнованые методы, для более удобной работы в будущем.
+ 
+Для того, чтобы найти "точки роста" для оптимизации я воспользовался профилирование RubyProf
 
 Вот какие проблемы удалось найти и решить
