+++ lib/worker.rb
@@ -1,0 +1,161 @@
+require_relative './user'
+require 'date'
+require 'json'
+
+class Worker
+  def initialize(file_path)
+    @file_path = file_path
+    @users = []
+    @sessions = []
+  end
+
+  def run
+    parsing_file
+    report_build
+  end
+
+  private
+
+  def report_build
+    # Отчёт в json
+    #   - Сколько всего юзеров +
+    #   - Сколько всего уникальных браузеров +
+    #   - Сколько всего сессий +
+    #   - Перечислить уникальные браузеры в алфавитном порядке через запятую и капсом +
+    #
+    #   - По каждому пользователю
+    #     - сколько всего сессий +
+    #     - сколько всего времени +
+    #     - самая длинная сессия +
+    #     - браузеры через запятую +
+    #     - Хоть раз использовал IE? +
+    #     - Всегда использовал только Хром? +
+    #     - даты сессий в порядке убывания через запятую +
+
+    report = {}
+
+    report[:totalUsers] = @users.count
+
+    # Подсчёт количества уникальных браузеров
+    report['uniqueBrowsersCount'] = find_count_unique_browsers
+
+    report['totalSessions'] = @sessions.count
+
+    report['allBrowsers'] = find_all_browsers
+
+    # Статистика по пользователям
+    users_objects = generate_users_objects
+
+    report['usersStats'] = {}
+
+    # Собираем количество сессий по пользователям
+    collect_stats_from_users(report, users_objects) do |user|
+      { 'sessionsCount' => user.sessions.count }
+    end
+
+    # Собираем количество времени по пользователям
+    collect_stats_from_users(report, users_objects) do |user|
+      { 'totalTime' => user.sessions.map {|s| s['time']}.map {|t| t.to_i}.sum.to_s + ' min.' }
+    end
+
+    # Выбираем самую длинную сессию пользователя
+    collect_stats_from_users(report, users_objects) do |user|
+      { 'longestSession' => user.sessions.map {|s| s['time']}.map {|t| t.to_i}.max.to_s + ' min.' }
+    end
+
+    # Браузеры пользователя через запятую
+    collect_stats_from_users(report, users_objects) do |user|
+      { 'browsers' => user.sessions.map {|s| s['browser']}.map {|b| b.upcase}.sort.join(', ') }
+    end
+
+    # Хоть раз использовал IE?
+    collect_stats_from_users(report, users_objects) do |user|
+      { 'usedIE' => user.sessions.map{|s| s['browser']}.any? { |b| b.upcase =~ /INTERNET EXPLORER/ } }
+    end
+
+    # Всегда использовал только Chrome?
+    collect_stats_from_users(report, users_objects) do |user|
+      { 'alwaysUsedChrome' => user.sessions.map{|s| s['browser']}.all? { |b| b.upcase =~ /CHROME/ } }
+    end
+
+    # Даты сессий через запятую в обратном порядке в формате iso8601
+    collect_stats_from_users(report, users_objects) do |user|
+      { 'dates' => user.sessions.map{|s| s['date']}.map {|d| Date.parse(d)}.sort.reverse.map { |d| d.iso8601 } }
+    end
+
+    File.write('result.json', "#{report.to_json}\n")
+  end
+
+  def generate_users_objects
+    users_objects = []
+
+    @users.each do |user|
+      attributes = user
+      user_sessions = @sessions.select { |session| session['user_id'] == user['id'] }
+      user_object = User.new(attributes: attributes, sessions: user_sessions)
+      users_objects = users_objects + [user_object]
+    end
+
+    users_objects
+  end
+
+  def find_all_browsers
+    @sessions.map { |s| s['browser'] }
+             .map { |b| b.upcase }
+             .sort
+             .uniq
+             .join(',')
+  end
+
+  def find_count_unique_browsers
+    uniqueBrowsers = []
+
+    @sessions.each do |session|
+      browser = session['browser']
+      uniqueBrowsers += [browser] if uniqueBrowsers.all? { |b| b != browser }
+    end
+
+    uniqueBrowsers.count
+  end
+
+  def parsing_file
+    file_lines = File.read(@file_path).split("\n")
+
+    file_lines.each do |line|
+      cols = line.split(',')
+      @users = @users + [parse_user(line)] if cols[0] == 'user'
+      @sessions = @sessions + [parse_session(line)] if cols[0] == 'session'
+    end
+  end
+
+  def parse_user(user)
+    fields = user.split(',')
+
+    {
+      'id' => fields[1],
+      'first_name' => fields[2],
+      'last_name' => fields[3],
+      'age' => fields[4],
+    }
+  end
+
+  def parse_session(session)
+    fields = session.split(',')
+
+    {
+      'user_id' => fields[1],
+      'session_id' => fields[2],
+      'browser' => fields[3],
+      'time' => fields[4],
+      'date' => fields[5],
+    }
+  end
+
+  def collect_stats_from_users(report, users_objects, &block)
+    users_objects.each do |user|
+      user_key = "#{user.attributes['first_name']}" + ' ' + "#{user.attributes['last_name']}"
+      report['usersStats'][user_key] ||= {}
+      report['usersStats'][user_key] = report['usersStats'][user_key].merge(block.call(user))
+    end
+  end
+end
